fix: Comprehensive bug fixes from multi-agent audit
- fractal_embedder: Fix cosine similarity (add normalization)
- fractal_storage: Fix parameter name (storage_path)
- memory_graph_trainer: Remove duplicate class (665 lines)
- learning_opportunity_manager: Fix SQL query to match schema
- learning_opportunity: Change priority to float
- self_reasoning_engine: Fix recursion depth off-by-one
- knowledge_graph: Fix logic error in _get_recent_updates_for_domain
- knowledge_graph: Remove duplicate check in update_node_weights
- knowledge_graph: Move safe_json_loads after imports
- analyzer_core: Remove dead code (unbound methods)
- curiosity_engine: Populate curiosity_history list
- component_initializer: Fix FractalStorage parameter name

+++ cogniflex/learning/learning_opportunity.py
@@ -6,7 +6,7 @@ from typing import Any, Dict, List, Optional
 class LearningOpportunity:
     concept: str
     opportunity_type: str
-    priority: int
+    priority: float
     domain: str
     evidence: List[Any]
     suggested_actions: List[Any]
